+++ hydromt_delwaq/workflows/emissions.py
@@ -76,12 +76,13 @@ def gridlength_gridwidth(ds):
 
 
 def emission_raster(
-    da,
-    ds_like,
-    method="average",
-    fillna_method="nearest",
-    fillna_value=0.0,
-    area_division=False,
+    da: xr.DataArray,
+    ds_like: xr.Dataset,
+    method: str = "average",
+    fillna_method: str = "nearest",
+    classnumber: int | float = 0.0,
+    fillna_value: int | float = 0.0,
+    area_division: bool = False,
 ):
     """Return emission map.
 
@@ -94,11 +95,13 @@ def emission_raster(
         DataArray containing emission map.
     ds_like : xarray.DataArray
         Dataset at model resolution.
-    method : str {'average', 'nearest', 'mode'}
+    method : str {'average', 'nearest', 'mode', 'classfraction', 'classarea'}
         Method for resampling.
+    classnumber : int | float
+        Class number used for resampling methods 'classfraction' or 'classarea'.
     fillna_method : str {'nearest', 'zero', 'value'}
         Method to fill NaN values.
-    fillna_value : float
+    fillna_value : int | float
         If fillna_method is set to 'value', NaNs in the emission maps will be replaced
         by this value.
     area_division : boolean
@@ -126,14 +129,52 @@ def emission_raster(
             nodata = -999.0
         da.raster.set_nodata(nodata)
 
+    if method == "classfraction" or method == "classarea":
+        area_division = False
+        # Final nodata for classfraction and classarea is -9999
+        nodata = -9999.0
+
     if area_division:
         da_area = gridarea(da)
         da = da / da_area
 
     logger.info(f"Deriving {da.name} using {method} resampling (nodata={nodata}).")
-    # da = da.astype(np.float32)
 
-    da_out = da.raster.reproject_like(ds_like, method=method)
+    if method == "classfraction" or method == "classarea":
+        da = da.astype("int32")
+        # return 1 for classnumber, NULL for all other classes
+        da_boolean = da.where(da.values == classnumber, 0) / classnumber
+        # High resolution raster
+        if da.raster.res[0] < ds_like.raster.res[0]:
+            # convert to area
+            da_area = da_boolean * gridarea(da)
+            # Reproject using sum
+            da_out = da_area.raster.reproject_like(ds_like, method="sum")
+            if method == "classfraction":
+                # convert to fraction using area at model resolution
+                da_area_model = gridarea(ds_like)
+                da_out = da_out / da_area_model
+                # Max at 1 (avoid rounding errors)
+                da_out = da_out.where(da_out.values <= 1.0, 1.0)
+        else:
+            # Reproject using average
+            da_out = da_boolean.raster.reproject_like(ds_like, method="average")
+            if method == "classarea":
+                # convert to area using area at model resolution
+                da_area_model = gridarea(ds_like)
+                da_out = da_out * da_area_model
+        # Change dtype now to float32
+        da_out = da_out.astype("float32")
+
+    # method is 'average', 'nearest' or 'mode'
+    elif method in ["average", "nearest", "mode"]:
+        da_out = da.raster.reproject_like(ds_like, method=method)
+    else:
+        raise ValueError(
+            f"Resampling method {method} not recognized. "
+            "Choose from 'average', 'nearest', 'mode', 'classfraction', 'classarea'."
+        )
+
     if area_division:
         da_area = gridarea(da_out)
         da_out = da_out * da_area